+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - WOM-591: Display the current and end index of the loaded track table in the 'Track' tab.
 - Bug fixes:
     - WOM-597: Check that the path in the track table file input field is a real file and not a directory to prevent an error message when pressing 'Load Track Table File'.
+    - WOM-576: Add an "Attribute" menu to display all attributes as graphs or logs in their own window.
 
 ## 0.5.2